리팩토링 진행 전 submain 합치기
