--- src/pages/train/components/train/ExplainTrainOption.js
@@ -1,10 +1,0 @@
-function ExplainTrainOption({title, explain, id}) {
-  return (
-    <>
-      <label htmlFor={id}>{title}</label>
-      <p className="small-explain">{explain}</p>
-    </>
-  )
-}
-
-export default ExplainTrainOption
