+++ src/pages/train/utils/InputHandler.js
