+++ src/pages/train/utils/StationToggleHandler.js
